Move Period to domain/period.py
