+++ savoten/domain/__init__.py
@@ -1,2 +1,3 @@
-from .event import Event, Period  # noqa: F401
+from .event import Event  # noqa: F401
+from .period import Period  # noqa: F401
 from .user import User  # noqa: F401
